# backend/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Service Tracking Platform API",
        "description": "REST API for the Service Tracking Platform with AI powered service request classification and technician assignment.\n        \n        Features:\n        - JWT-based authentication and authorization\n        - AI Smart Request Creation\n        - AI Technician Recommendation\n        - Status Transition Management\n        - Real-time Assignment Workflow\n        ",
        "contact": {
            "name": "Mileta Lvova",
            "email": "mileta279@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/locations/": {
            "get": {
                "tags": ["Locations"],
                "summary": "Get all locations",
                "description": "Endpoint to get all locations",
                "responses": {
                    "200": {
                        "description": "List of locations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of locations"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "post": {
                "tags": ["Locations"],
                "summary": "Creates a location",
                "description": "Endpoint to create a location",
                "responses": {
                    "201": {
                        "description": "location created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Location created"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Location"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationInput"
                            }
                        }
                    }
                }
            }
        },
        "/api/locations/{id}": {
            "get": {
                "tags": ["Locations"],
                "summary": "Get locations by ID",
                "description": "Endpoint to get details of a specific location by its ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Location ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of a location retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Locations details"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Location"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "put": {
                "tags": ["Locations"],
                "summary": "Updates a location",
                "description": "Endpoint to update a location",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Location ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Location updated"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Location"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound",
                        "description": "Not Found"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationUpdateInput"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Locations"],
                "summary": "Delete a location",
                "description": "Endpoint to delete a location",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Location ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Location deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound",
                        "description": "Not Found"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Authorization header using the Bearer scheme. \n                Enter ONLY your JWT token in the input below (without 'Bearer ' prefix).\n                \n                Example: \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\""
            }
        },
        "schemas": {
...
            "Location": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 1
                                    }
                                }
                            },
                            "address": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    }
                                }
                            },
                            "city": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "New York"
                                    }
                                }
                            },
                            "state": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "NY"
                                    }
                                }
                            },
                            "zipCode": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "10001"
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized - Authentication required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - Insufficient permissions",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "JWT": []
        }
    ]
}